Collect more of GPML's output, update documentation
Under some of the modes of calling GPML's gp(), there can be some
elements of the returned Octave value list that are empty. When called
in prediction mode without test targets, it gives an output list of
length six, where we want elements 0, 1, 2, 3, and 5 rather than an
output list of five where we want elements 0 through 4. When called
in training mode, you have to tell it the nargout is 2 rather than 1
like usual due to lines 137--138 of gp.m:

    if nargout<=1
      [post nlZ] = feval(inf{:}, hyp, mean, cov, lik, x, y); dnlZ = {};

This causes you to not get dnlZ if you use

    feval(gp, in, 1);

like usual; we have to instead use

    feval(gp, in, 2);

to get all the output we want.