Support Octave 4.4.1
The Octave version in the Cosmic repos (Ubuntu 18.10), and available
for install on Mac via Homebrew, is Octave 4.4.1. Using that version
requires using C++11 standard and requires a change to the embedding
approach. I couldn't get the approach in the Octave docs working, so
I tried our old octave_main() approach for Octave < 4.2.0. That threw
a warning (at runtime) that its arguments were going to be ignored,
so I tried it without any arguments and it works fine. This does
result in a deprecated warning at compile time, but no other issues.
I was able to deal with all the other deprecation that ocurred from
Octave >= 4.4.0 just fine.