Ensure Octave path prints to Rcout, improve tests
If we really want to test whether Octave is fully set up OK when the
package is attached, we need to check out the Octave load path. The
issue there is Octave's path() function prints to cout, not Rcout, so
capture.output() won't catch it. So, we redirect cout to Rcout. We may
want to change this whole approach later, but it works for now.

So, that allows us not only to make sure that Octave's load path gets
set correctly when the package is attached in our tests, but also
expand our test coverage to the .print_path() function. I also noticed
that in our tests we don't actually cover the verbose branch in
.embed_octave() for the initial embedding, so I made a minor change to
.onAttach() to fix that, and added a test to make sure that converting
from Rcpp::NumericMatrix to an Octave Matrix doesn't cause problems.