Add support for composite mean/cov functions
The initial setup only supported simple mean and covariance functions.
I add an additional type conversion function, list_to_cell, on the C++
side to allow for composite mean and covariance functions. I also
update the type conversion function map_to_list to allow for Octave to
pass function handles to R, as I found out that some of the inference
methods return a function handle for POST rather than a matrix.

I also take care of a couple of bugs here, notably that GPML's gp()
apparently needs to be called from the GPML directory for some
specifications, even if Octave's load path includes all of the sub-
folders of the GPML directory (this is true even if calling gp()
straight from Octave -- i.e., it's not specific to how we call it).
For now, our workaround is to just set the working directory in the
R-facing gp() function to the GPML directory, then reset it back.