Change string_vector conversion, update tests
Some versions of the Octave C++ API don't have a string_vector
constructor from an std::vector<std::string>, but it looks like all
versions can do it from an std::list<std::string>, so we move to that
conversion safely for all versions.

I updated the tests to make sure the type and dimensions of all the
output was correct, but not to test for numerical accuracy to runs from
my Linux machine; it may be that there will be slight numerical
differences between systems. I need to investigate this further.