Move to GPML v3.6
The current version of GMPL (4.2) has a bug in its covPref function,
which is causing problems for applications for the lab. We're moving
back to version 3.6, which makes all the gp() calculations correct.
However, this does cause R to complain quite a bit that 'a function
in gpml/util/unwrap.m shadows a core library function', so I'll have
to look into that more later.