Fix Octave path
Moving to GMPL v 3.6 has changed the subdirectories we need to add
to the Octave path. It also caused an Octave warning about unwrap
masking an Octave core library function, so I go ahead and redirect
both cout and cerr as soon as Octave is attached to hide these
things from the user.